starter/requirements.txt
flask-moment
flask-wtf
python-dotenv==0.10.3
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
Jinja2==2.10.1
psycopg2-binary==2.8.2
pytz==2019.1
SQLAlchemy==1.3.4
pylint==2.3.1
